Add Load/Save_EFER entry/exit controls for VM profile
Setting these entry and exit controls is recommended by Intel SDM Vol.
3C, section 31.2.1.

When the load IA32_EFER VM entry control is set, the IA32_EFER is loaded
with the value from the corresponding guest VMCS field on guest entry.

When the store IA32_EFER VM exit control is set, the IA32_EFER is stored
to the corresponding guest VMCS field on guest exit.

When the load IA32_EFER VM exit control is set, the IA32_EFER is loaded
with the value from the corresponding host VMCS field on guest exit.