skpolicy: Set PDPTE/PDE page map flag
Use PDPTE or PDE page mappings depending on the alignment (1G or 2M) of
a given memory region.

The writable and executable flags are set to the values of the
respective memory region attributes only, if the entry maps a page of
the corresponding size. Otherwise the entries are marked as writable
and executable since the final paging structure entry that maps the page
frame determines the effective flag values.