Calculate PDE target address for large page mapping
If a page directory entry (PDE) maps a 2 MB page frame, the address is
calculated depending on the physical address of the memory region.

This allows for mapping of a 1 GB continguous memory region using PDE
mappings.

Note: The page table generation code is restricted to one table of each
      paging structure level (PML4, PDPT, PD and PT).