Enable 64bit VM support
VMs still boot in 32bit mode, but are be able to transfer to 64bit.
Also implement using bzImage as linux source format, which gives us
a unified 32bit entry point for both kernel types.

There doesn't seem to be a non-FPU AMD64 implementation out there, Linux
devs got lazy, so this enables FPU, SSE, SSE2, FXSTOR.
The OS can (and likely will) rely on their presence.
(but we don't support it yet)

TODO:
This currently has a hack in that it bypasses a number of instructions
after the 32bit entry point. This is because the kernel first reloads
the segment selectors, which we set, but never back with a descriptor table.
Accessing this non-existant table leads to failure.
Unfortunately that offset to the entry point isn't stable, so this needs
fixing.