libmugen: Add Mugen.Generators package
The Generators.Run procedure uses the Command_Line package to read the
output directory and the path of the policy file. It then parses the
policy and passes the output directory and XML data on to the given
generator procedure.

This can be used by the generator tools to easily run their own
generators.