muspecgen: Implement kernel C header file generation
The VMXON region address is determined by the physical address of the
memory region called 'kernel_0|vmxon'. For the VMCS region address the
memory region called 'tau0|vmcs' is used. It is assumed, that VMCS
regions are placed continuously in physical memory and that the VMCS
region of Tau0 is at the beginnig.