muspecgen: Add Spec.Utils package
The package contains a generic function which computes the numeric value
for a given list of bitfield nodes. The generic must be instantiated
with an enumeration type encompassing all fields and an array mapping
the bitfields to their position.

This can be used to convert, e.g. Pin-Base VM-Execution controls to their
numeric representation.