Do not set ref addresses when adding mem regions
Only set the destination address of paging structure entries that map a
page frame. Set the address to zero if it is a reference to another
table entry.

The references can be adjusted once the physical addresses of all paging
structures are fixed. This is a prerequisite for supporting arbitrary
numbers of paging structures.