muvalidate: Add For_Each_Match procedure
For each element specified by 'Source_XPath', the procedure tries to
find a match in the nodes specified by 'Ref_XPath' using the given
'Match' function. The message is appended to the log message. If no
match is found, an exception with the message returned by the 'Error'
function raised.