mugenspec: Add Vector_Route_Type to interrupts template
The new Vector_Route_Type allows to associate a host vector with a
destination subject id and a subject vector number. This decouples the
host vector, which is delivered to the kernel when a device raises an
IRQ, and the number of the vector, which is injected into the subject.

Note: This commit temporarily breaks the build to ease review of the
      change set. The necessary adaptations on the kernel-side follow
	  in the next commit.