Mugenzp: Only declare subject memory as usable
Memory marked as usable in the e820 map can potentially be used for DMA
by Linux. As for now only the extra memory can be identity-mapped, we
simply deny the Linux kernel to use other regions. This restriction will
be removed once VT-d is implemented.

NOTE: With this change Linux can properly use PCI hardware such as the
      e1000e ethernet controller. This was verified by acquiring an IP
      address using dhclient and pinging the codelabs.ch server.