Add Context_Table_Type to VTd.Tables package
This type represents a DMAR context table as specified by the Intel VT-d
specification, section 9.3.

The Add_Entry procedure adds an entry to a DMAR context table for a
device, function pair with given domain identifier and a pointer to the
base of second-level paging entries.

Context table entries are created with 39-bit AGAW (3-level second-level
page table) and 00b translation type. DID is restricted to 255 security
domains.