Add initial Paging.Layouts package
The package contains a memory layout type that is a collection of
logical to physical memory mappings managed in several levels of paging
structures.

It is based on the Paging.Memory package. The new data type and
operations will support the creation of 3-level page tables as needed
for VT-d second-level address translation structures.