Implement Paging.Layouts.Add_Memory_Region procedure
The procedure adds a memory region with the specified attributes to the
given memory layout. All necessary paging structure entries on the
different paging levels are created accordingly.

The code is more compact in comparison to the
Paging.Memory.Add_Memory_Region procedure since only the first level
structures must be treated separately.