Libpaging: Do not set bit 7 of EPT PTEs
Only set the flag in directory and not in PT entries to avoid issues
with VT-d hardware. As per specification bit 7 in EPT PTEs (see Intel
SDM Vol. 3C, table 28-1) as well as VT-d SLPTEs (see Intel
Virtualization Technology for Directed I/O, table 9-41) is supposed to
be ignored by hardware. Unfortunately for Ivy Bridge chipsets this is
not actually true: if the bit is set the DMAR engine logs a fault with
reason 16#0c# meaning 'Non-zero reserved field in a second-level paging
entry is set'.

NOTE: It is noteworthy that earlier versions of the VT-d spec (revision
      1.3, February 2011) are *very* specific about this field being
      ignored: 'Hardware always ignores the programming of this field in
      leaf page-table entries corresponding to 4KB pages'.