Libmutools: Add Get_Active_CPU_Count utility function
The function returns the number of CPUs that are active in a given
system policy. Active means that the CPU is part of the scheduling plan
and is used to execute subjects. The number is determined by counting
the cpu child elements of the first major frame. The validator must
guarantee that each major frame specifies the same amount of CPU
elements.