Implement Elfcheck.Run
The procedure reads a given system policy and a binary object file and
validates that properties of certain sections match the properties of
memory regions specified in the policy.

The checked regions are currently hard-coded to 'kernel_text',
'kernel_data', 'kernel_ro' and 'kernel_bss', so the tool is only
useful to validate the kernel for now.