Mucfgvalidate: Disable device domain mapping check
Remove Device_Domains.Memory_Mapping_Address_Equality check as it is too
restrictive to be correct in all possible use-cases. For example, a
subject could interact with a device without using DMA. In this case,
enforcing virtual address equality in the mappings is not required.

The check procedure in Mucfgcheck is not deleted as it could be used as
optional check when needed.

Auditors: ken