Mucfgexpand: Implement major frame barriers expander
This new expansion step adds barrier configurations to each major frame.
A barrier config consists of a number of barriers, each barrier
specifying its size (i.e. the number of CPUs synchronizing on that
barrier).

The implementation works as follows: Insert the exit time of each minor
frame into an ordered list. The exit time of a minor frame is defined as
the number of ticks from the beginning of the major frame until the
minor frame ends. After all minor frames are inserted into the list,
chronologically traverse the ordered list and add a barrier whenever
multiple, consecutive exit times are identical. The size of the barrier
is given by the number of list elements with the same exit time.