Libmucfgcheck: Adjust PCI LSI IRQ range check
Level signaled interrupts of PCI devices can span the entire range of
IRQs handled by the I/O APIC, including the range of legacy IRQs
(0..15). Additionally adjust the upper bound of the LSI range to account
for the fact that in general I/O APICs support 24 IRQs starting at zero.

Accordingly, the Physical_IRQ_Constraints_PCI_LSI validator as changed
to allow the definition of PCI devices with LSIs in the range of 0..23.