Mucfgmerge: Do not validate hardware XML files on load
Hardware files are not validated against the XML schema when read from
file. Instead missing elements are added as necessary prior to merging.
This makes sure that the section structure is correct after the merge
operation.

Note: This does not enable the creation of invalid system policies since
the merged XML file is validated against the format source schema.