Extend subject device BDF allocator
Handle PCI multi-function devices when expanding subject device BDFs.
All logical devices that are part of a physical PCI multi-function
device are assigned the same logical bus and device number. If a subject
device has a device number assigned then this number will also be used
for other devices that are part of the same PCI multi-function device.