Merge Libmusinfo project with Mugensinfo
The current Libmusinfo library contains code to generate the sinfo page.
What we actually need is a SPARK 2014 library to access the sinfo
information from within a component. Therefore merge the files in
Libmusinfo with the Mugensinfo project in order to make room in the
namespace for the upcoming components/libmusinfo library.

Note: This changes the license of the imported Libmusinfo code from BSD
      back to GPLv3.