Mugenspec: Prove two range checks in Skp.IOMMU spec
These appeared just recently with the new toolchain. The reason is that
the exact Size of the IOMMU_[1|2]_Type, if not specified by the user, is
decided by the GNAT backend and can't be easily back-propagated to
GNATprove ([Q404-048]).

Therefore, explicitly specify the Size of IOMMU_[1|2]_Type. It must be
calculated by the tool as the order of IOTLB invalidate / fault
registers is not deterministic.