Fix inconsistency in Section_Infos
Up to now, the decision whether a binary section is represented by a
<file> or a <fill> node was based on two variables in the respective
Section_Info: Write_To_File and Fill.  This allowed inconsistencies,
such as both Write_To_File and Fill being set to False.

Now, the decision is solely based on Write_To_File.  The field Fill
has been replaced by a field Fill_Pattern, which allows to specify the
byte value the section is to be filled with.

The BSS section is now represented by a <fill> node with fill pattern
16#00#.