Expand channels in stage 1
This is required to have all possible switch events in place before
subject runnability checks and CPU ID assignment. Otherwise subjects
with only channel associated events are detected as non-runnable.

Note that the global and subject channel elements are removed in stage 2
after channel checks have been performed in the pre-check procedure of
that stage. Having the checks in stage 1 does not work because
components/libraries are also expanded in stage 1.