Copy memory type in Handle_Loaders
If the newly created *_dst memory is set to 'subject' type, it will be
announced as usable to Linux, making it a possible DMA32 zone region.
This only works if the region is also part of a device domain.

Avoid this by copying the 'type' attribute of the original memory
region.