Add Get_Binary_Section function to Run package
The function returns True and the requested binary section if present in
the binary given by descriptor.

If the section is not found, and the section is not marked optional in
the section info, an exception is raised. False is returned if the
section is marked as optional in the section info record.