tools: Add Solo5 Manifest to cspec conversion script
The script processes a Solo5 unikernel and a source component spec. The
Solo5 manifest [1] is extracted from the binary using the
'solo5-elftool', which is expected to be available in the PATH (see
[2]).

It adds required memory regions with the correct access rights for all
ELF segments. A ram region with the size defined by the '--ram'
parameter is inserted (default: 512 MB, same as Solo5/Hvt). Furthermore
a corresponding reader/writer channel pair for each NET_BASIC device
declared in the manifest is added.  The optional '--bootparams'
parameter can be used to specify unikernel boot parameters.

[1] - https://github.com/Solo5/solo5/issues/372
[2] - https://github.com/Solo5/solo5/pull/433