Add Mirage/Solo5 Net policy creation script
The compose.py script takes a template policy and merges in the
component specification which is expected to be called
'component_unikernel.xml' and located in the same directory.
The following steps are performed for each channel required by the
component:

- Add physical channel
- Add subject component mapping to unikernel subject
- Add logical channel to NIC Linux subject

Additionally, the subject boot parameters are set according to the
components 'bootparams' config value. Furthermore, the NIC Linux boot
parameters are extended with the 'unikernel_iface' which specifies all
unikernel muennet interfaces as a CSV list.