integ: Rework SSH connect retry logic
The ConnectionAttempts and ConnectTimeout SSH client options did not
have the desired effect for the QEMU target because the QEMU user
networking hostfwd mechanism uses a buffered connection, i.e. from the
view of the ssh client, the connection immediately succeeds even though
the SSH daemon in the guest is not ready yet.

The buffered packets are forwarded as soon as the guest networking
becomes ready. This creates a race: if the SSH daemon is not ready at
this point, the SSH client receives a connection reset and does not
retry (because the connection already succeeded).

Therefore, drop aforementioned options in favour of a new execute_n
function which retries a given command N times, with an intermediate
sleep. Use the execute_n function for initial connections to lnx1 and
lnx2 guests.