Mark writable initrd regions as RAM
Let Linux reclaim/-use initrd memory regions which are writable.
Otherwise the chunk of memory remains unused for the lifetime of the
Linux subject. Additionally, Linux does not need to copy the initramfs
into RAM (again) since the location where it is mapped is usable as is
and not marked as reserved via the e820 map.