Add schedulingPartitionType to policy schema
It will be used to specify scheduling partitions, which consist of a
number of scheduling groups. The difference between format source and
formats A&B is that scheduling groups do not have an ID and CPU
attribute in policy format source. The attributes are optional for now
to enable an iterative implementation.