Add scheduling partition config to spec
For each scheduling partition, the new config specifies the following:

- The number of scheduling groups in the partition
- The scheduling groups (by ID) that are part of the partition

The scheduling group index range is chosen to start from zero since it
will be used in conjunction with the group activity bitmap. This avoids
having to subtract or add 1 when using it to access the bitmap.

Note: The initially active scheduling group is the first entry in the
      scheduling group ID map.