Fix bug in mutools-xml_templates
The bug occurred if a config variable "foo" was referenced in a template
with parameter "foo" on toplevel. Then, compile_template would first add
the local "foo" to its database of local variables and later on think that
value="$foo" refers to that local variable and rename the reference.
A cyclic reference followed.
Now, the values of parameters are marked as "locked" and never renamed.