Only short-circuit handover events without action
The conversion of target events without an action to invalid target
events is now limited to handovers. For all other events, a
corresponding target event with No_Action is generated. This will lead
to the "full" event processing cycle, meaning the target event will be
marked as pending and on the target side it will be consumed as an event
with action = No_Action. This way events without actions can be used to
wakeup potentially sleeping subjects without requiring an inject
interrupt target action and the associated additional processing.