Tools: Adjust virtualAddress domain
The number of bits that can be used for virtual addresses is limited.
That limit turns out to be below 48 bit (sometimes 36 bit,
sometimes 39 bit). The domain for virtual addresses used by
mucfg[c]vresalloc for linuxes needs to take this into account.
Otherwise, an error ocurrs (in ioremap_cache called by Muennet).