scripts: Add --disable-reset param to solo5-muen-gencspec.py
If the command line argument is given, the corresponding component
config value will be set to false. By default it is true, which means a
reset event with the appropriate logical name will be automatically
added. This event, if present, will be triggered by Solo5 whenever it
reaches platform_exit().