Mugenspec: Cleanup VMX_Types package
Move implementation details to private part of spec and provide separate
implementation for functions. This cleans up the public part of the spec
and is also a preparatory step for an upcoming function which requires
an implementation in the package body. GNATtest will trip up if the
functions are directly provided as generic instantiations.