Mugendts: Add kernel image to usable memory again
The Linux kernel seems to sometimes access its own memory via virtual
addresses (seen to happen for the runtime alternatives code patching),
which fails if those virtual addresses are not part of the kernel's
page tables.

That it continued to work even after the referenced commit was most
likely due to an overlap.  Because the size of the kernel image was
set too low in the policy, parts of the image's virtual address
space (about 2.5 MiB) actually fell into the subject memory, which
directly followed the kernel image in the virtual as well as the
physical memory map.

51c13f47dcc7 ("Mugendts: Fix memory type checks for Linux memory nodes")