Mugends: Add support to extract and pad files
This extracts parts of files that are specified with `offset` in the
policy, and pads files that are smaller than the memory section in
which they are to be placed.  Exceptions are raised if offsets or
files are too large.

Depending on the platform, the padding could later be optimized so it
does not actually have to be transmitted as part of the file.