Mugends: Optimize file padding and fill regions for u-boot deployment
Instead of actually loading a padded file or a file with a certain fill
pattern, it can be more efficient to just instruct a boot loader that
supports it to fill a certain memory region with a given pattern.

For bootgen there isn't a way to fill a memory region and for gdb and
xsct I've not found an efficient way to achieve it.