Mugenspec: Base SMMU configuration on device domains
Instead of basing the configuration on single subjects and their
complete page tables, use the device domains and their separate page
tables.

The subject ID is removed from the controller configuration as there is
no direct relation with subjects anymore.

The VMID is currently set to the ID of the device domain.  Since the
kernel enables private VMID namespacing for the SMMU, broadcast TLB
maintenance operations with a matching VMID should be ignored (however,
this option is only a hint, i.e. the operations are not "required" to
affect the TLB entries but they still could).