Libmucfgcheck: Add a check for the alignment of Linux images on ARM
As documented, the ARM64 Linux boot process expects the base address of
the kernel image to be 2MB aligned (we get a "[Firmware Bug]" otherwise).

This check locates Linux subjects that have a device tree memory node
assigned and checks the alignment of the virtual address of the subject
binary nodes.