Mugendts: Mark kernel image as reserved memory
We don't map this area into the SMMU but we still have to define it as
usable memory so the kernel creates page tables as it accesses some data
via virtual addresses. Reserving the area prevents the kernel from using
it as regular memory.  Note that a kernel patch is necessary to avoid
that the "initmem" section between code and data is later marked as usable
again.