Mugends: Abort boot if any tftpboot command fails
By default, the boot just continues if any command fails, later causing
errors whose cause might not immediately be obvious.  Now the boot
is aborted immediately after a failed `tftpboot` command via the `panic`
command.  This either halts or reboots the system depending on whether
`CONFIG_PANIC_HANG` is enabled in the U-Boot config, which already is
the case for the Xilinx ZynqMP default config.  No message is passed to
`panic`, which causes an empty line to be printed.  The output looks
something like this:

    TFTP from server x.x.x.x; out IP address is y.y.y.y
    Filename 'kernel.bin-kernel_text.part'.
    Load address: 0x0
    Loading: *
    TFTP error: 'File not found' (1)
    Not retrying...

    ### ERROR ### Please RESET the board ###