Mucfgucode: Fix multi processor signature handling
Summary:
If a given microcode update is suited for multiple processor signatures,
the processor signature in the header is set to the lowest value of all
supported signatures, not the one requested via iucode-tool -s
parameter.

In order to avoid consulting the extended signature table, the offset of
which depends on the dynamic data size value in the header, adjust the
primary processor signature to the signature requested and update the
header checksum accordingly.

All supported processor IDs are still stored in the extended signature
table, so no information is lost.

Test Plan:
Applied an MCU suitable for multiple processor signatures where the
requested processor signature was not in the signature header field by
default (0x90675 vs. 0x90672).

Reviewers: ken

Reviewed By: ken

Differential Revision: https://dev.codelabs.ch/D1145