Mucheckstack: Collect .ci files also relative to project dir
When building with bob, GPR files of dependencies generally don't
specify an object dir.  There is usually a library dir configured but
since that varies, it's not so easy to put the .ci files there.
However, the GPR files themselves are always in the same location. So
if we place the .ci in a sibling directory, we can locate them here
relative to the project dir. We first check if a .ci file can be found
there and fallback to the object dir for the current component itself.